Polish code style, comments and fix typos

+++ inloop/grading/models.py
@@ -9,12 +9,8 @@ from django.dispatch import receiver
 from inloop.solutions.models import Solution
 
 
-def get_upload_path(obj, filename):
-    """
-    Return an upload file path.
-
-    All Plagiarism Test zip files will be stored in a shared directory.
-    """
+def zipfile_upload_path(obj, filename):
+    """Return upload file paths for the PlagiarismTest.zip_file field."""
     return "plagiarism_tests/{created_at}/{filename}".format_map({
         "created_at": obj.created_at,
         "filename": filename,
@@ -24,28 +20,24 @@ def get_upload_path(obj, filename):
 class PlagiarismTest(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     command = models.TextField(default="", help_text="Command that was used to perform the test")
-    zip_file = models.FileField(upload_to=get_upload_path, null=True)
+    zip_file = models.FileField(upload_to=zipfile_upload_path, null=True)
 
     def __str__(self):
         return "Plagiarism test #{}".format(self.id)
 
     @property
     def all_detected_plagiarisms(self):
-        """
-        Get all detected plagiarisms to this Test.
-        """
-        return DetectedPlagiarism.objects.filter(test=self, veto=False)
+        """Get all detected plagiarisms for this test."""
+        return self.detectedplagiarism_set.filter(veto=False)
 
 
 @receiver(post_delete, sender=PlagiarismTest, dispatch_uid="delete_plagiarism_file")
 def auto_delete_file_on_delete(sender, instance, **kwargs):
     """
-    Deletes the zip file
-    when its corresponding PlagiarismTest object is deleted.
+    Removes the zip file when its corresponding PlagiarismTest object is deleted.
     """
-    if instance.zip_file:
-        if os.path.isfile(instance.zip_file.path):
-            os.remove(instance.zip_file.path)
+    if instance.zip_file and os.path.isfile(instance.zip_file.path):
+        os.remove(instance.zip_file.path)
 
 
 class DetectedPlagiarism(models.Model):
@@ -71,13 +63,11 @@ def get_ripoff_tasks_for_user(user):
 
 
 def save_plagiarism_set(plagiarism_set, result_dir):
-    """Save the detected plagiarisms to the database and assign the zipped output"""
+    """Save the detected plagiarisms and zip file to the database."""
     path_to_zip = make_archive(result_dir, "zip", result_dir)
     with open(path_to_zip, 'rb') as zip_data:
         zip_file = SimpleUploadedFile("jplag_test.zip", zip_data.read())
-    test = PlagiarismTest.objects.create(
-        zip_file=zip_file
-    )
+    test = PlagiarismTest.objects.create(zip_file=zip_file)
     DetectedPlagiarism.objects.bulk_create([
         DetectedPlagiarism(test=test, solution=solution) for solution in plagiarism_set
     ])
